requirements.txt
alabaster==0.7.12
asgiref==3.2.10
attrs==19.3.0
Babel==2.8.0
bleach==3.1.5
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
CommonMark==0.5.4
datarum==0.3.0
decorator==4.4.2
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.8
djangorestframework==3.11.0
docutils==0.16
entrypoints==0.3
gunicorn==20.0.4
idna==2.10
imagesize==1.2.0
insegel==1.1.0
ipython-genutils==0.2.0
Jinja2==2.11.2
jsonschema==3.2.0
jupyter-core==4.6.3
jupyter-sphinx-theme==0.0.6
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.7
nbsphinx==0.7.1
packaging==20.4
pandocfilters==1.4.2
parse==1.15.0
psycopg2==2.8.5
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
pytz==2020.1
pywin32==228
recommonmark==0.4.0
requests==2.24.0
six==1.15.0
snowballstemmer==2.0.0
Sphinx==3.1.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.3.1
testpath==0.4.4
traitlets==4.3.3
urllib3==1.25.9
webencodings==0.5.1
whitenoise==5.1.0
